Detect if server started from client is no longer running

+++ src/base/system.cpp
@@ -4149,15 +4149,32 @@ int kill_process(PROCESS process)
 {
 #if defined(CONF_FAMILY_WINDOWS)
 	BOOL success = TerminateProcess(process, 0);
-	if(success)
+	BOOL is_alive = is_process_alive(process);
+	if(success || !is_alive)
 	{
 		CloseHandle(process);
+		return true;
 	}
-	return success;
+	return false;
 #elif defined(CONF_FAMILY_UNIX)
+	if(!is_process_alive(process))
+		return true;
 	int status;
 	kill(process, SIGTERM);
 	return waitpid(process, &status, 0) != -1;
+#endif
+}
+
+bool is_process_alive(PROCESS process)
+{
+	if(process == INVALID_PROCESS)
+		return false;
+#if defined(CONF_FAMILY_WINDOWS)
+	DWORD exit_code;
+	GetExitCodeProcess(process, &exit_code);
+	return exit_code == STILL_ACTIVE;
+#else
+	return waitpid(process, nullptr, WNOHANG) == 0;
 #endif
 }
